Change `fromTwitterFuture` to accept a call-by-name Future (#127)
Currently the usage of fromTwitterFuture expects a com.twitter.util.Future to
be wrapped in a Task, amongst other things, in order to defer the execution of
the Future until the effect is executed. However it can be simplified by using
a call-by-name argument and then wrapping it in a Task within the implementation.

In order to keep things backwards compatible we are keeping the current
interface and marking it deprecated, and moving the implementation to a shared
private function that is just called from both entrypoints.
We should subsume them in a future release.